fix(decode): 质量值频率表元素计数用 uint32_t, 修掉 uint8_t 回绕导致的堆越界
fastq_actuator.cpp 与 sam_actuator.cpp 两处同型写法:

    uint8_t qualFreqArrLen = qualFreqSrcLen / sizeof(uint16_t);
    uint16_t* qualFreqArray = new uint16_t[qualFreqArrLen];
    qualFreqCoder->decode_line((uint8_t*)qualFreqArray, qualFreqSrcLen, ...);

数组按截断后的元素数分配, decode_line 却按未截断的 qualFreqSrcLen 字节写入。
质量值字母表超过 127 个符号时(qualFreqSrcLen > 510)计数回绕, 分配变小而写入不变,
直接堆越界。Illumina 常见字母表约 38 个符号不会踩到, 非 Illumina 编码或
含异常字符的数据会。

排查 FASTQ 偶发损坏时顺带发现, 经核实本数据(38 个符号 -> srcLen 152, /2 = 76)
不触发它, 所以这不是那个缺陷的修复。

回归: 带参考/无参考/管道/区域查询/流式全部通过, 压缩产物 md5 未变(97e353e7...)。

+++ src/fastq_actuator.cpp
@@ -1462,7 +1462,12 @@ int32_t FastqCodecActuator::initDecoder(RoughIOBlock* outputBlock) {
         auto qualFreqCoder = std::make_unique<coder_bwt_cm>(&qualFreqIo);
 
         uint32_t qualFreqSrcLen = streamsMeta[1]["srclen"].asUInt();
-        uint8_t qualFreqArrLen = qualFreqSrcLen / sizeof(uint16_t);
+        /*
+         * 元素个数必须用 uint32_t: 数组按这个数分配, 而 decode_line 按未截断的
+         * qualFreqSrcLen 字节写入。原来是 uint8_t, 一旦质量值字母表超过 127 个符号
+         * (qualFreqSrcLen > 510) 计数就回绕, 分配变小而写入不变, 直接堆越界。
+         */
+        uint32_t qualFreqArrLen = qualFreqSrcLen / sizeof(uint16_t);
         uint16_t* qualFreqArray = new uint16_t[qualFreqArrLen];
         uint32_t qualFreq = qualFreqCoder->decode_line((uint8_t*)qualFreqArray, qualFreqSrcLen, UINT8_MAX, false);
         if (qualFreq != qualFreqSrcLen) {
